modal move to delayed.js
