+++ eds/blocks/centered-content-switcher/centered-content-switcher.js
@@ -1,4 +1,5 @@
 import { createOptimizedPicture } from '../../scripts/aem.js';
+import { decorateModal } from '../../scripts/delayed.js';
 import {
   calciteButton,
   ul,
@@ -151,6 +152,11 @@ export default function decorate(block) {
       anchor.appendChild(playButton);
       anchor.parentElement.parentElement.removeChild(anchor.parentElement);
       tab.children[0].appendChild(anchor);
+      playButton.addEventListener('click', (evt) => {
+        evt.preventDefault();
+        playButton.setAttribute('tabindex', '0');
+        decorateModal(anchor.href, playButton);
+      });
     } else {
       const button = calciteButton({
         'icon-end': 'arrowRight',
