+++ eds/blocks/hero/hero.js
@@ -1,79 +1,80 @@
 import { createAutoplayedVideo } from '../../scripts/scripts.js';
 import { calciteButton } from '../../scripts/dom-helpers.js';
+import { decorateModal } from '../../scripts/delayed.js';
 
-let lastfocusBtn;
-const toggleLoader = () => {
-  const loader = document.querySelector('.web-dev-loader');
-  if (loader) {
-    loader.classList.toggle('visible');
-  }
-};
+// let lastfocusBtn;
+// const toggleLoader = () => {
+//   const loader = document.querySelector('.web-dev-loader');
+//   if (loader) {
+//     loader.classList.toggle('visible');
+//   }
+// };
 
-const removeModal = (modal) => {
-  const videoIframeWrapper = document.querySelector('.video-iframe-wrapper');
-  if (videoIframeWrapper) {
-    videoIframeWrapper.remove();
-  }
+// const removeModal = (modal) => {
+//   const videoIframeWrapper = document.querySelector('.video-iframe-wrapper');
+//   if (videoIframeWrapper) {
+//     videoIframeWrapper.remove();
+//   }
 
-  ['style', 'tabindex', 'aria-hidden'].forEach((attr) => document.body.removeAttribute(attr));
-  modal.remove();
-  lastfocusBtn.focus();
-};
+//   ['style', 'tabindex', 'aria-hidden'].forEach((attr) => document.body.removeAttribute(attr));
+//   modal.remove();
+//   lastfocusBtn.focus();
+// };
 
-const handleEscKeyPress = (event) => {
-  if (event.key === 'Escape') {
-    removeModal(document.querySelector('.co3-modal'));
-  } else {
-    const co3ModalContainer = document.querySelector('.co3-modal-container');
-    if (co3ModalContainer) {
-      co3ModalContainer.focus();
-    }
-  }
-};
+// const handleEscKeyPress = (event) => {
+//   if (event.key === 'Escape') {
+//     removeModal(document.querySelector('.co3-modal'));
+//   } else {
+//     const co3ModalContainer = document.querySelector('.co3-modal-container');
+//     if (co3ModalContainer) {
+//       co3ModalContainer.focus();
+//     }
+//   }
+// };
 
 // decorate modal
-function decorateModal(videoLink) {
-  const closeButton = Object.assign(document.createElement('calcite-icon'), {
-    className: 'co3-modal-container calcite-icon',
-    icon: 'x',
-    scale: 'm',
-    tabIndex: 0,
-    ariaLabel: 'close modal',
-    ariaHidden: 'false',
-  });
+// function decorateModal(videoLink) {
+//   const closeButton = Object.assign(document.createElement('calcite-icon'), {
+//     className: 'co3-modal-container calcite-icon',
+//     icon: 'x',
+//     scale: 'm',
+//     tabIndex: 0,
+//     ariaLabel: 'close modal',
+//     ariaHidden: 'false',
+//   });
 
-  const iframe = document.createElement('iframe');
-  const modalContainer = document.createElement('div');
-  const modal = document.createElement('div');
-  iframe.classList.add('co3-modal', 'iframe');
-  iframe.setAttribute('src', videoLink);
-  modalContainer.classList.add('co3-modal-container');
-  modalContainer.setAttribute('tabindex', '0');
-  modalContainer.appendChild(iframe);
-  modalContainer.appendChild(closeButton);
-  modal.classList.add('co3-modal', 'calcite-mode-dark');
-  modal.appendChild(modalContainer);
-  document.body.setAttribute('tabindex', '-1');
-  document.body.setAttribute('aria-hidden', 'true');
-  document.body.appendChild(modal);
-  document.addEventListener('keydown', handleEscKeyPress);
-  closeButton.addEventListener('click', () => removeModal(modal));
-  closeButton.addEventListener('keydown', (event) => {
-    if (event.key === 'Enter' || event.key === ' ' || event.key === 'Escape') {
-      removeModal(modal);
-    }
-  });
-  modal.addEventListener('click', (event) => {
-    if (event.target === modal) {
-      removeModal(modal);
-    }
-  });
+//   const iframe = document.createElement('iframe');
+//   const modalContainer = document.createElement('div');
+//   const modal = document.createElement('div');
+//   iframe.classList.add('co3-modal', 'iframe');
+//   iframe.setAttribute('src', videoLink);
+//   modalContainer.classList.add('co3-modal-container');
+//   modalContainer.setAttribute('tabindex', '0');
+//   modalContainer.appendChild(iframe);
+//   modalContainer.appendChild(closeButton);
+//   modal.classList.add('co3-modal', 'calcite-mode-dark');
+//   modal.appendChild(modalContainer);
+//   document.body.setAttribute('tabindex', '-1');
+//   document.body.setAttribute('aria-hidden', 'true');
+//   document.body.appendChild(modal);
+//   document.addEventListener('keydown', handleEscKeyPress);
+//   closeButton.addEventListener('click', () => removeModal(modal));
+//   closeButton.addEventListener('keydown', (event) => {
+//     if (event.key === 'Enter' || event.key === ' ' || event.key === 'Escape') {
+//       removeModal(modal);
+//     }
+//   });
+//   modal.addEventListener('click', (event) => {
+//     if (event.target === modal) {
+//       removeModal(modal);
+//     }
+//   });
 
-  iframe.addEventListener('load', () => {
-    toggleLoader();
-    modalContainer.focus();
-  });
-}
+//   iframe.addEventListener('load', () => {
+//     toggleLoader();
+//     modalContainer.focus();
+//   });
+// }
 
 export default function decorate(block) {
   const newChildren = [...block.children].map((entry) => {
@@ -144,10 +145,7 @@ export default function decorate(block) {
       calciteBtn.addEventListener('click', (evt) => {
         evt.preventDefault();
         calciteBtn.setAttribute('tabindex', '0');
-        lastfocusBtn = calciteBtn;
-
-        toggleLoader();
-        decorateModal(videoLink.href);
+        decorateModal(videoLink.href, calciteBtn);
       });
     }
   }
