+++ eds/styles/styles.css
@@ -546,3 +546,49 @@ main {
 .padding-trailer-6 {
   padding-block-end: 9rem;
 }
+
+/* Modal */
+body[aria-hidden="true"]{
+  overflow: hidden;
+}
+
+.co3-modal {
+  background: var(--esri-ui-opacity80-inverse);
+  block-size: 100%;
+  inline-size: 100vw;
+  inset-block-start: 50%;
+  inset-inline-start: 50%;
+  position: fixed;
+  transform: translate(-50%, -50%);
+  z-index: 3001;
+}
+
+.co3-modal-container {
+  aspect-ratio: 1.7778;
+  inline-size: 85%;
+  inset-block-start: 50%;
+  inset-inline-start: 50%;
+  position: absolute;
+  transform: translate(-50%, -50%);
+}
+
+.co3-modal iframe {
+  block-size: 100%;
+  inline-size: 100%;
+}
+
+ .co3-modal-container calcite-icon {
+  background: var(--calcite-ui-text-1);
+  block-size: 50px;
+  border-radius: 100%;
+  inline-size: 50px;
+  inset-block-start: 0;
+  inset-inline-start: 100%;
+  z-index: 3002;
+}
+
+.co3-modal-container calcite-icon:is(:hover, :focus-within, :active) {
+  background: var(--calcite-ui-brand);
+  color: var(--calcite-ui-text-1);
+  cursor: pointer;
+}
